+++ lib/llm/providers/google/stream_parser.rb
@@ -150,6 +150,7 @@ class LLM::Google
         fn.arguments = call["args"]
         fn.tracer = @stream.extra[:tracer]
         fn.model = @stream.extra[:model]
+        fn.guard = @stream.extra[:guard]
       end
       [fn, (registered ? nil : fn.unavailable)]
     end